py-polars/docs/requirements-docs.txt
docutils
hypothesis
numpy
pandas
pyarrow

sphinx==8.1.3

# Third-party Sphinx extensions
autodocsumm==0.2.14
numpydoc==1.8.0
pydata-sphinx-theme==0.16.0
sphinx-autosummary-accessors==2023.4.0
sphinx-copybutton==0.5.2
sphinx-design==0.6.1
sphinx-favicon==1.0.1
sphinx-reredirects==0.1.5
sphinx-toolbox==3.8.1
railroad-diagrams==3.0.1

livereload==2.7.0
